Shortened/balanced duck game
Increased player and duckling speeds; reduced quack circle strength,
radius, and cool-down. increased duckling wander strength and wander
angle range, decreased passive duckling mother follow, strengthened
rivers, reduced game time for perfect score, tested & balanced. can
usually get 9 or 10 ducklings by 1:45

+++ Assets/Scripts/FlockingScripts/FlockingAI.cs
@@ -67,7 +67,7 @@ public class FlockingAI : MonoBehaviour
     public float leaderInfluenceDistance;
 
     public bool followsMother = false;
-    public int motherInfluenceDistance = 40;
+    public float motherInfluenceDistance = 40;
     public GameObject mother;
     public float motherWeight = 1f; //multiplier for strength of mother influence
